Use sidebar radio for navigation
